Added choose images modal
Clicking choose in search modal results now shows book covers to choose
from if the first one is not desired. It takes it from the searchResults.
Looping through the items in the object and creating a div > img
to append to the modal-content. Uses some styling from the book display